Add DTOs and apply them to Service and Controller Layers

+++ UserManagement API/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace UserManagement_API.Interfaces
     public interface IUserRepository
     {
         Task<IEnumerable<User>> GetAllAsync();
-        Task<User> GetByIdAsync(int id);
+        Task<User> GetByIdAsync(long id);
         Task AddAsync(User user);
         Task UpdateAsync(User user);
         Task DeleteAsync(User user);
