+++ UserManagement API/Repositories/UserRepository.cs
@@ -30,7 +30,7 @@ namespace UserManagement_API.Repositories
             return await _context.Users.ToListAsync();
         }
 
-        public async Task<User> GetByIdAsync(int id)
+        public async Task<User> GetByIdAsync(long id)
         {
            return await _context.Users.FindAsync(id);
         }
